Made all of the callbacks optional by creating __weak versions
Decided that it would be best to change the strategy for callbacks as a callback may be added to the core code and we would not want to have to touch every board to implement.  With the new design we could add a new feature to the core and add a callback any only boards that decide to take on that feature would need to change.