Fix to ensure that the echo frame is immediately sent back to the host
Since we use the RX IRQ for adding messages to the queue there may be instances where the echo frame would be pushed behind a normal RX frame.  To keep using the RX IRQ we need to ensure that the echo frame is sent in the next queue check.  Pushing it to the front ensures that even if an IRQ comes in before the queue check the echo frame will the the first out.