Add tasks for the handling of the queues
After reading more on FreeRTOS this is the recommended method for using tasks to handle queues.  By polling the queue it prevents the system from going to idle while there is no data to process.  This allows the system to properly preempt tasks if there is new queue data waiting.